remove plugin service discovery test integration module as the existing arc test integration module covers that, added plugin-specific service discovery test to arc test integration test module.

+++ test/mock/plugins-sync/src/plugins/myplugin.js
@@ -15,7 +15,7 @@ module.exports = {
       callback()
     },
     end: function ({ arc, inventory, services }, callback) {
-      fs.unlinkSync(file, 'syncplugin.test')
+      if (fs.existsSync(file)) fs.unlinkSync(file, 'syncplugin.test')
       callback()
     }
   }
